refactor: remove extraneous html, restructure project
