+++ src/actor-sheet/actor-sheet-npc.component.html
@@ -1,11 +1,11 @@
-<form class="{{cssClass}} {{actor.type}} flexcol" autocomplete="off">
+<form class="{{ cssClass }} {{ actor.type }} flexcol" autocomplete="off">
   {{!-- Sheet Header --}}
   <header class="sheet-header">
     <img
       class="profile-img"
-      src="{{actor.img}}"
+      src="{{ actor.img }}"
       data-edit="img"
-      title="{{actor.name}}"
+      title="{{ actor.name }}"
       height="100"
       width="100"
     />
@@ -14,7 +14,7 @@
         <input
           name="name"
           type="text"
-          value="{{actor.name}}"
+          value="{{ actor.name }}"
           placeholder="Name"
         />
       </h1>
@@ -31,14 +31,14 @@
             <input
               type="text"
               name="data.health.value"
-              value="{{data.health.value}}"
+              value="{{ data.health.value }}"
               data-dtype="Number"
             />
             <span> / </span>
             <input
               type="text"
               name="data.health.max"
-              value="{{data.health.max}}"
+              value="{{ data.health.max }}"
               data-dtype="Number"
             />
           </div>
@@ -50,14 +50,14 @@
             <input
               type="text"
               name="data.power.value"
-              value="{{data.power.value}}"
+              value="{{ data.power.value }}"
               data-dtype="Number"
             />
             <span> / </span>
             <input
               type="text"
               name="data.power.max"
-              value="{{data.power.max}}"
+              value="{{ data.power.max }}"
               data-dtype="Number"
             />
           </div>
@@ -69,7 +69,7 @@
             <input
               type="text"
               name="data.cr"
-              value="{{data.cr}}"
+              value="{{ data.cr }}"
               data-dtype="Number"
             />
             <span> / </span>
@@ -77,7 +77,7 @@
               type="text"
               disabled="true"
               name="data.xp"
-              value="{{data.xp}}"
+              value="{{ data.xp }}"
               data-dtype="Number"
             />
           </div>
@@ -106,12 +106,12 @@
 
     {{!-- Owned Items Tab --}}
     <div class="tab items" data-group="primary" data-tab="items">
-      {{> "systems/cryptomancer/templates/actor/parts/actor-items.html"}}
+      {{> "systems/cryptomancer/actor-sheet/parts/actor-items.html"}}
     </div>
 
     {{!-- Active Effects Tab --}}
     <div class="tab effects flexcol" data-group="primary" data-tab="effects">
-      {{> "systems/cryptomancer/templates/actor/parts/actor-effects.html"}}
+      {{> "systems/cryptomancer/actor-sheet/parts/actor-effects.html"}}
     </div>
   </section>
 </form>
