+++ src/cryptomancer.scss
@@ -1,0 +1,23 @@
+// Add custom fonts by visiting and search https://fonts.google.com
+// @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap");
+
+// Import utilities.
+// @import "utils/typography";
+// @import "utils/colors";
+// @import "utils/mixins";
+// @import "utils/variables";
+
+/* Global styles */
+// @import "global/window";
+// @import "global/grid";
+// @import "global/flex";
+
+@import "skill-check/roll";
+
+/* Styles limited to cryptomancer sheets */
+.cryptomancer {
+  // @import "components/forms";
+  // @import "components/resource";
+  // @import "components/items";
+  // @import "components/effects";
+}
