+++ src/cryptomancer.ts
@@ -2,9 +2,9 @@
 import { DropData } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/foundry.js/clientDocumentMixin";
 
 // Import document classes.
-import { CryptomancerActor } from "./documents/actor.js";
+import { CryptomancerActor } from "./actor/actor.component.js";
 // Import sheet classes.
-import { CryptomancerActorSheet } from "./sheets/actor-sheet.js";
+import { CryptomancerActorSheet } from "./actor-sheet/actor-sheet.component.js";
 // Import helper/utility classes and constants.
 import { preloadHandlebarsTemplates } from "./helpers/templates.js";
 // import { BOILERPLATE } from "./helpers/config.mjs";
