--- src/module/helpers/templates.ts
@@ -1,15 +1,0 @@
-/**
- * Define a set of template paths to pre-load
- * Pre-loaded templates are compiled and cached for fast access when rendering
- * @return {Promise}
- */
-export const preloadHandlebarsTemplates = async function () {
-  return loadTemplates([
-    // Actor partials.
-    "systems/cryptomancer/templates/actor/parts/actor-features.html",
-    "systems/cryptomancer/templates/actor/parts/actor-items.html",
-    "systems/cryptomancer/templates/actor/parts/actor-spells.html",
-    "systems/cryptomancer/templates/actor/parts/actor-effects.html",
-    "systems/cryptomancer/templates/actor/parts/actor-attribute.html",
-  ]);
-};
