+++ src/skill-check/cryptomancer-roll.html
@@ -1,0 +1,15 @@
+<div class="skill-check">
+  {{log rolls}}
+  <div class="attribute">
+    <h3>{{attributeName}} - {{skillName}}</h3>
+  </div>
+  <div class="difficulty">{{difficulty}}</div>
+  <div class="result">{{checkResult}}</div>
+  <ol class="dice">
+    {{#each rolls.attribute}}
+    <li class="crypto-die attr result-{{this.value}} {{result}}"></li>
+    {{/each}} {{#each rolls.fate}}
+    <li class="crypto-die fate result-{{this.value}} {{result}}"></li>
+    {{/each}}
+  </ol>
+</div>
