+++ src/skill-check/skill-check.enum.ts
